+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+* Added conditional CSS transform of single color stop linear gradient ([#4518](https://github.com/evanw/esbuild/issues/4518))
+
+    ```css
+    /* Original code */
+    a { background: linear-gradient(red); }
+
+    /* New output */
+    a { background: linear-gradient(red, red); }
+    ```
+
 * Allow `es2026` as a target in `tsconfig.json`
 
     TypeScript is [adding `es2026`](https://github.com/microsoft/TypeScript/issues/63704) as a compilation target, so esbuild now supports this in the `target` field of `tsconfig.json` files, such as in the following configuration file: